refactor: recombine constants, storage, types
