--- src/utils/global/storage.ts
@@ -1,20 +1,0 @@
-import * as Types from "./types";
-
-// Global settings
-export const globalSettings = storage.defineItem<Types.Settings>("local:globalSettings", {
-  version: 1,
-  fallback: {
-    global: true,
-    plugins: true,
-    themes: true,
-    snippets: true,
-
-    themeFlavour: "mocha",
-    themeAccent: "mauve",
-    themeLogo: "schooltape-rainbow",
-
-    userSnippets: {},
-    urls: ["https://help.schoolbox.com.au"],
-    needsRefresh: false,
-  },
-});
