--- src/utils/plugins/constants.ts
@@ -1,44 +1,0 @@
-import { PluginId, PluginInfo } from "./types";
-
-export const PLUGIN_INFO: Record<PluginId, PluginInfo> = {
-  subheader: {
-    name: "Subheader Revamp",
-    description: "Adds a clock and current period info to the subheader",
-    order: 0,
-  },
-  scrollSegments: {
-    name: "Scroll Segments",
-    description: "Segments the Schoolbox page into scrollable sections",
-    order: 1,
-  },
-  scrollPeriod: {
-    name: "Scroll Period",
-    description: "Scrolls to the current period on the timetable",
-    order: 2,
-  },
-  modernIcons: {
-    name: "Modern Icons",
-    description: "Modernise the icons across Schoolbox",
-    order: 3,
-  },
-  tabTitle: {
-    name: "Better Tab Titles",
-    description: "Improves the tab titles for easier navigation",
-    order: 4,
-  },
-  homepageSwitcher: {
-    name: "Homepage Switcher",
-    description: "The logo will switch to existing Schoolbox homepage when available",
-    order: 5,
-  },
-  timetableLabels: {
-    name: "Timetable Labels",
-    description: "Labels the day of the week from numbers to the actual day",
-    order: 6,
-  },
-  legacyTimetable: {
-    name: "Legacy Timetable",
-    description: "Moves the timetable to its own row",
-    order: 7,
-  },
-};
